added media settings for cloudinary

+++ set_env.py
@@ -20,9 +20,11 @@ def read_data(file_path: str | Path) -> dict:
 
 
 def write_data(file_path: str | Path, data: dict):
-    with open(file_path, "w") as file:
-        for key, value in data.items():
-            file.write(f"{key}='{value}'\n")
+    file_path = Path(file_path)
+    if file_path.is_file():
+        with open(file_path, "w") as file:
+            for key, value in data.items():
+                file.write(f"{key}='{value}'\n")
 
 
 if __name__ == "__main__":
